Вынес логику создания нового пользователя, логгирование и так далее из тестовых модулей
