+++ tests/functional/post_v1_account/test_post_v1_account.py
@@ -1,54 +1,3 @@
-from datetime import datetime
-from collections import namedtuple
-
-import pytest
-
-from heplers.account_helper import AccountHelper
-from restclient.configuration import Configuration as MailhogConfiguration
-from restclient.configuration import Configuration as DmApiConfiguration
-from services.dm_api_account import DMApiAccount
-from services.api_mailhog import MailHogApi
-import structlog
-
-structlog.configure(
-    processors=[
-        structlog.processors.JSONRenderer(
-            indent=4,
-            ensure_ascii=True,
-            # sort_keys=True
-        )
-    ]
-)
-
-@pytest.fixture
-def mailhog_api():
-    mailhog_configuration = MailhogConfiguration(host='http://5.63.153.31:5025')
-    mailhog_client = MailHogApi(configuration=mailhog_configuration)
-    return mailhog_client
-
-
-@pytest.fixture
-def account_api():
-    dm_api_configuration = DmApiConfiguration(host='http://5.63.153.31:5051', disable_log=False)
-    account = DMApiAccount(configuration=dm_api_configuration)
-    return account
-
-@pytest.fixture
-def account_helper(account_api, mailhog_api):
-    account_helper = AccountHelper(dm_account_api=account_api, mailhog=mailhog_api)
-    return account_helper
-
-@pytest.fixture
-def prepare_user():
-    now = datetime.now()
-    data = now.strftime("%d_%m_%Y_%H_%M_%S")
-    login = f'golovan_{data}'
-    password = '112233'
-    email = f'{login}@mail.ru'
-    User = namedtuple("User", ["login", "password", "email"])
-    user = User(login=login, password=password, email=email)
-    return user
-
 def test_post_v1_account(account_helper, prepare_user):
     login = prepare_user.login
     password = prepare_user.password
