+++ tests/functional/put_v1_account_token/test_put_v1_account_token.py
@@ -1,32 +1,8 @@
-from heplers.account_helper import AccountHelper
-from restclient.configuration import Configuration as MailhogConfiguration
-from restclient.configuration import Configuration as DmApiConfiguration
-from services.dm_api_account import DMApiAccount
-from services.api_mailhog import MailHogApi
-import structlog
 
-structlog.configure(
-    processors=[
-        structlog.processors.JSONRenderer(
-            indent=4,
-            ensure_ascii=True,
-            # sort_keys=True
-        )
-    ]
-)
-
-def test_put_v1_account_token():
-    dm_api_configuration = DmApiConfiguration(host='http://5.63.153.31:5051', disable_log=False)
-    mailhog_configuration = MailhogConfiguration(host='http://5.63.153.31:5025')
-
-    account = DMApiAccount(configuration=dm_api_configuration)
-    mailhog = MailHogApi(configuration=mailhog_configuration)
-
-    account_helper = AccountHelper(dm_account_api=account, mailhog=mailhog)
-
-    login = 'golovan181'
-    password = '112233'
-    email = f'{login}@mail.ru'
+def test_put_v1_account_token(account_helper, prepare_user):
+    login = prepare_user.login
+    password = prepare_user.password
+    email = prepare_user.email
 
     account_helper.register_new_user(login=login, password=password, email=email)
     account_helper.user_login(login=login, password=password)
